AP_HAL: fix signed and unsigned comparison warning
../../libraries/AP_HAL/examples/Printf/Printf.cpp:63:17: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
         if (ret != strlen(float_tests[i].result)) {
                 ^
...

ardupilot/modules/gtest/include/gtest/gtest.h:1448:16: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   if (expected == actual) {
                ^
And similar ones.

+++ libraries/AP_HAL/examples/Printf/Printf.cpp
@@ -60,7 +60,7 @@ static void test_printf(void)
                                 float_tests[i].result);
             failures++;
         }
-        if (ret != strlen(float_tests[i].result)) {
+        if (ret != (int)strlen(float_tests[i].result)) {
             hal.console->printf("Failed float_tests[%u] ret=%d/%d '%s' should be '%s'\n",
                                 (unsigned)i,
                                 ret, (int)strlen(float_tests[i].result),
